requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
discord.py==2.3.2
mistralai==0.0.7
